Incremental tree and layout loading is working, mostly.
Still a lot to do.  Had to temporarily disable styling and overview.  Look for TODOs and FIXMEs.
Firefox doesn't seem to be triggering the render after the tree initially loads (but does once the view is scrolled or zoomed)
Still needs database for the servlets
Added the json dependency to the pom.xml
Made a HashMap version of LayoutCladogram for the layout servlet (using node UUIDs for keys)